+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added "CDC/ATSDR Social Vulnerability Index 2020" variable to "CDC/ATSDR Social Vulnerability Index" algorithm based on census 2020, changed 2018 suffix of existing variable to 2010 (based on census year).
 - Added support for years 2018-2021 using census 2020 in "NAACCR Yost Quintile & Area-Based Social Measures Linkage Program" algorithm.
 - Added support for years 2018-2021 using census 2020 in "NAACCR Poverty Linkage Program" algorithm.
+- Added new "Breast Cancer Category" algorithm.
 
 **Changes in version 5.5**